Made progress in physics_mg.cpp
Decomposed nu_fission into prompt+delayed in two places.
When actually using the frequencies, there is a piece of code
that calls prn() and changes the random number sequence.
Consequently, all tests will fail after that change. So I'm
leaving that change for last.

Just discovered that mgxs.cpp has a function sample_fission_energy
that needs to be modified for delayed_nu_fission to use frequencies.
It uses delayed_nu_fission as modified in physics_mg. So I need
to define the full vector in physics_mg, pass it to sample_fission_energy
and then change sample_fission_energy in addition to changing every isntance
of sample_fission_energy elsewhere in OpenMC.

Note: I still haven't validated p.macro_xs_.inverse_velocity
or mesh_bin or freq_group or freq. I need to make sure those
still work outside of the if-statements they were modified in.

+++ src/particle.cpp
@@ -330,7 +330,7 @@ Particle::event_collide()
   if (settings::run_mode == RunMode::EIGENVALUE &&
       type_ == Particle::Type::neutron) {
     
-    double mesh_bin = -1;
+    int mesh_bin = -1;
     double nu_fission = macro_xs_.prompt_nu_fission;
     if (settings::precursor_frequency_on) {
       mesh_bin = simulation::frequency_mesh->get_bin(r());
